requirements.txt
# ============================================================================
# EDGE-RAG UNIFIED REQUIREMENTS
# ============================================================================
#
# Version: 4.0.0
# Author: Jan Nietzard
# Last Modified: 2026-04-23
#
# These are range-constrained requirements for development.
# For reproducible thesis evaluation, use requirements_frozen.txt (pip freeze).
#
# Installation:
#   pip install -r requirements.txt
#
# With spaCy NER:
#   pip install spacy
#   python -m spacy download en_core_web_sm
#
# ============================================================================

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# LangChain Framework
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0

# LangSmith (required by LangChain 0.3.x)
langsmith>=0.2.0,<1.0.0

# LangGraph — agentic controller
langgraph>=1.0.0,<2.0.0
langgraph-checkpoint>=2.1.0,<5.0.0

# Data Processing
pydantic>=2.5.0,<3.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Numerical Computing
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Networking
requests>=2.31.0
httpx>=0.25.0

# Progress Bars
tqdm>=4.65.0

# Typing Extensions
typing-extensions>=4.9.0

# ============================================================================
# VECTOR STORE (LanceDB — embedded)
# ============================================================================
lancedb>=0.6.0,<0.8.0
pyarrow>=12.0.0

# ============================================================================
# GRAPH DATABASE
# ============================================================================

# KuzuDB (required — NetworkX fallback removed in v4.0.0)
kuzu>=0.3.0

# ============================================================================
# PDF PROCESSING
# ============================================================================
pypdf>=4.0.0
pdf2image>=1.16.0
python-docx>=1.1.0

# ============================================================================
# BENCHMARK DEPENDENCIES
# ============================================================================

# HuggingFace Datasets — HotpotQA, 2WikiMultihop, StrategyQA
datasets>=2.14.0

# ============================================================================
# OPTIONAL: Enhanced Entity Extraction (NER)
# ============================================================================

# spaCy — traditional NER (recommended for edge)
# Installation: pip install spacy && python -m spacy download en_core_web_sm
# spacy>=3.5.0

# GLiNER — zero-shot NER (larger, more flexible)
# gliner>=0.2.0

# ============================================================================
# OPTIONAL: Thesis Evaluation and Visualization
# ============================================================================

# Data Analysis
# pandas>=2.0.0

# Visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0

# ============================================================================
# VERSION COMPATIBILITY NOTES
# ============================================================================
#
# langchain 0.3.x requires langchain-core 0.3.x (not 1.x).
# langgraph 1.0.x accepts langchain-core>=0.3.0.
# networkx was removed in v4.0.0 — KuzuDB is the only supported graph backend.
#
# ============================================================================
